# src/client/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DUCC Website API",
    "description": "API documentation for the Durham University Canoe Club website.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie used for authentication."
      }
    }
  },
  "paths": {
...
    "/api/files": {
        "get": {
            "summary": "List files",
            "description": "Retrieves a paginated list of uploaded files, with optional filtering.",
            "tags": ["Files"],
            "parameters": [
                { "name": "page", "in": "query", "description": "Page number to retrieve.", "schema": { "type": "integer" } },
                { "name": "limit", "in": "query", "description": "Number of files per page.", "schema": { "type": "integer" } },
                { "name": "search", "in": "query", "description": "Search term to filter files by title.", "schema": { "type": "string" } },
                { "name": "categoryId", "in": "query", "description": "Filter files by category ID.", "schema": { "type": "integer" } }
            ],
            "responses": {
                "200": { 
                  "description": "List of files",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "files": { "type": "array", "items": { "type": "object" }, "description": "Array of file objects." },
                          "totalPages": { "type": "integer", "description": "Total number of pages." },
                          "currentPage": { "type": "integer", "description": "Current page number." },
                          "totalFiles": { "type": "integer", "description": "Total number of files matching criteria." }
                        }
                      }
                    }
                  }
                }
            }
        },
        "post": {
            "summary": "Upload files",
            "description": "Uploads one or more files to the system with a title and category.",
            "tags": ["Files"],
            "security": [{ "cookieAuth": [] }],
            "requestBody": {
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "files": { "type": "array", "items": { "type": "string", "format": "binary" }, "description": "The files to upload." },
                                "title": { "type": "string", "description": "A title for the file upload." },
                                "categoryId": { "type": "integer", "description": "The category ID for the files." }
                            }
                        }
                    }
                }
            },
            "responses": {
                "201": { "description": "Files uploaded" }
            }
        }
    },
    "/api/files/{id}": {
        "put": {
            "summary": "Update file metadata",
            "description": "Updates file metadata such as title, category, or visibility.",
            "tags": ["Files"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The ID of the file.", "schema": { "type": "integer" } }],
            "requestBody": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "title": { "type": "string", "description": "New title for the file." },
                                "categoryId": { "type": "integer", "description": "New category ID." },
                                "visibility": { "type": "string", "description": "Visibility level (e.g., 'public', 'private')." }
                            }
                        }
                    }
                }
            },
            "responses": { "200": { "description": "File updated" } }
        },
        "delete": {
            "summary": "Delete file",
            "description": "Permanently deletes a file from the server.",
            "tags": ["Files"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The ID of the file to delete.", "schema": { "type": "integer" } }],
            "responses": { "200": { "description": "File deleted" } }
        }
    },
    "/api/files/{id}/download": {
        "get": {
            "summary": "Download file",
            "description": "Serves the file content for download or inline viewing.",
            "tags": ["Files"],
            "parameters": [
                { "name": "id", "in": "path", "required": true, "description": "The ID of the file.", "schema": { "type": "integer" } },
                { "name": "view", "in": "query", "schema": { "type": "boolean" }, "description": "If true, attempts to display the file inline instead of downloading." }
            ],
            "responses": { "200": { "description": "File content" } }
        }
    },
    "/api/file-categories": {
        "get": {
            "summary": "List file categories",
            "description": "Retrieves all categories used for organizing uploaded files.",
            "tags": ["Files"],
            "responses": { 
              "200": { 
                "description": "List of categories",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "array",
                      "items": { "type": "object", "properties": { "id": { "type": "integer", "description": "Category ID." }, "name": { "type": "string", "description": "Category name." } } }
                    }
                  }
                }
              } 
            }
        },
        "post": {
            "summary": "Create file category",
            "description": "Creates a new category for organizing files.",
            "tags": ["Files"],
            "security": [{ "cookieAuth": [] }],
            "requestBody": {
                "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string", "description": "Name of the category." } } } } }
            },
            "responses": { "200": { "description": "Category created" } }
        }
    },
    "/api/file-categories/{id}": {
        "put": {
            "summary": "Update file category",
            "description": "Updates the name of an existing file category.",
            "tags": ["Files"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The ID of the category.", "schema": { "type": "integer" } }],
            "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string", "description": "New name for the category." } } } } } },
            "responses": { "200": { "description": "Category updated" } }
        },
        "delete": {
            "summary": "Delete file category",
            "description": "Deletes an existing file category.",
            "tags": ["Files"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The ID of the category to delete.", "schema": { "type": "integer" } }],
            "responses": { "200": { "description": "Category deleted" } }
        }
    }
  }
}